+++ core/src/main/scala/com/databricks/labs/remorph/parsers/intermediate/commands.scala
@@ -87,3 +87,9 @@ case class WriteStreamOperationStart(
     foreach_batch: Option[StreamingForeachFunction])
     extends LeafNode
     with Command
+
+
+case class CreateVariable(name: String, dataType: DataType, defaultExpr: Option[Expression], replace: Boolean)
+  extends Command {}
+
+case class SetVariable(name: String, dataType: Option[DataType], expr: Option[Expression]) extends Command {}
