+++ core/src/main/scala/com/databricks/labs/remorph/parsers/intermediate/unresolved.scala
@@ -4,7 +4,6 @@ case class UnresolvedRelation(inputText: String) extends LeafNode {
   override def output: Seq[Attribute] = Seq.empty
 }
 
-
 case class UnresolvedStatement(inputText: String) extends Statement {}
 
 case class UnresolvedExpression(inputText: String) extends LeafExpression {
@@ -20,4 +19,3 @@ case class UnresolvedCatalog(inputText: String) extends Catalog {
   override def output: Seq[Attribute] = Seq.empty
   override def children: Seq[LogicalPlan] = Seq.empty
 }
-
