+++ core/src/test/scala/com/databricks/labs/remorph/parsers/snowflake/SnowflakeDDLBuilderSpec.scala
@@ -3,6 +3,7 @@ package com.databricks.labs.remorph.parsers.snowflake
 import com.databricks.labs.remorph.parsers.intermediate._
 import com.databricks.labs.remorph.parsers.snowflake.SnowflakeParser.{StringContext => _, _}
 import org.mockito.Mockito._
+import org.scalatest.Assertion
 import org.scalatest.matchers.should
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatestplus.mockito.MockitoSugar
@@ -15,7 +16,7 @@ class SnowflakeDDLBuilderSpec
 
   override protected def astBuilder: SnowflakeDDLBuilder = new SnowflakeDDLBuilder
 
-  private def example(query: String, expectedAst: Catalog): Unit = example(query, _.ddlCommand(), expectedAst)
+  private def example(query: String, expectedAst: Catalog): Assertion = example(query, _.ddlCommand(), expectedAst)
 
   "SnowflakeCommandBuilder" should {
     "translate Java UDF create command" in {
@@ -41,7 +42,7 @@ class SnowflakeDDLBuilderSpec
           name = "echo_varchar",
           returnType = VarCharType(None),
           parameters = Seq(FunctionParameter("x", VarCharType(None), None)),
-          JavaUDFInfo(
+          JavaRuntimeInfo(
             runtimeVersion = None,
             imports = Seq("@~/some-dir/some-lib.jar"),
             handler = "TestFunc.echoVarchar"),
@@ -73,7 +74,7 @@ class SnowflakeDDLBuilderSpec
           name = "py_udf",
           returnType = UnparsedType(),
           parameters = Seq(),
-          runtimeInfo = PythonUDFInfo(
+          runtimeInfo = PythonRuntimeInfo(
             runtimeVersion = Some("3.8"),
             packages = Seq("numpy", "pandas", "xgboost==1.5.0"),
             handler = "udf"),
@@ -105,7 +106,7 @@ class SnowflakeDDLBuilderSpec
           name = "js_factorial",
           returnType = DoubleType,
           parameters = Seq(FunctionParameter("d", DoubleType, None)),
-          runtimeInfo = JavascriptUDFInfo,
+          runtimeInfo = JavaScriptRuntimeInfo,
           acceptsNullParameters = false,
           comment = Some("Compute factorial using JavaScript"),
           body = javascriptCode))
@@ -133,7 +134,7 @@ class SnowflakeDDLBuilderSpec
           name = "echo_varchar",
           returnType = VarCharType(None),
           parameters = Seq(FunctionParameter("x", VarCharType(None), Some(Literal(string = Some("foo"))))),
-          runtimeInfo = ScalaUDFInfo(runtimeVersion = Some("2.12"), imports = Seq(), handler = "Echo.echoVarchar"),
+          runtimeInfo = ScalaRuntimeInfo(runtimeVersion = Some("2.12"), imports = Seq(), handler = "Echo.echoVarchar"),
           acceptsNullParameters = true,
           comment = None,
           body = scalaCode))
@@ -151,7 +152,7 @@ class SnowflakeDDLBuilderSpec
           parameters = Seq(
             FunctionParameter("a", DecimalType(None, None), None),
             FunctionParameter("b", DecimalType(None, None), None)),
-          runtimeInfo = SQLUDFInfo(memoizable = false),
+          runtimeInfo = SQLRuntimeInfo(memoizable = false),
           acceptsNullParameters = false,
           comment = Some("multiply two numbers"),
           body = "a * b"))
